a lot of cleanup, and introduction of a separate tMIDIKeyboard object. An instance of it is called MIDIKeyboard, and its the object that is controlled by USB MIDI devices and hosts.

+++ MOUSE_DUAL/MOUSE_DUAL/src/notestack.c
@@ -107,7 +107,7 @@ void tNoteStack_setCapacity(tNoteStack* const ns, uint8_t cap)
 		
 	for (int i = cap; i < MAX_NUM_NOTES; i++)
 	{
-		if (ns->notestack != -1)
+		if ((int)ns->notestack != -1)
 		{
 			ns->notestack[i] = -1;
 			ns->size -= 1;
